Create mipsCode
Sets strings that will be printed if input is too long, empty, or has an invalid base. These strings are "Input is too long.", "Input is empty." and "Invalid base-34 number." I used "asciiz" rather than "ascii" because the syscall needs to know where the string ends. The system call will keep on printing characters until it reaches a NUL-terminator.