Update mipsCode
The "addi" instructions adds the $0 register and 34 and stores the result in the $s0 register. The "addi" instructions adds the $0 register and 1 and stores the result in the $t4 register. The "addi" instructions adds the $0 register and 0 and stores the result in the $t5 register. The "addi" instructions adds the $0 register and 0 and stores the result in the $t6 register.