Update mipsCode
The branch skipSpaces allows code to not count space as a character or number. The "lb" instruction loads 0 into the $t1 register from $t0. The "addi" instruction adds $t0 and 1 and stores the result in the $t0 register. The "addi" instruction adds $t3 and 1 and stores the result in the $t3 register. The "beq" instruction implements the skipSpaces branch if the two registers are equal.