Update mipsCode
The "beq" instruction implements the emptyInputError if the $t1 register and 10 are equal. The "beq" instruction implements the emptyInputError if the $t1 register and $0 are equal. The "lb" instruction loads 0 into the $t1 register from the $t0 register. The "addition" instruction adds the $t0 register and 1 and stores the result in the $t0 register.