Update mipsCode
The "addition" instruction adds the $t3 register and 1 and stores the result in the $t3 register. The "beq" instruction implements the backToBeginning branch if $t1 and 10 are equal. The "beq" instruction implements the backToBeginning branch if $t1 and 32 are equal. The "one" instruction implements the showChars branch if $t1 and 32 are not equal.