Update mipsCode
The "lb" instruction loads 0 into $t1 from $t0. The "addi" instruction adds $t0 and 1 and stores the result in $t0.  The "addi" instruction adds $t3 and 1 and stores the result in $t3. The "beq" instruction implements the backToBeginning branch if $t1 and 10 are equal. The "beq" instruction implements the backToBeginning branch if $t1 and 0 are equal.