Update mipsCode
The "bne" instruction implements the invalidBaseError branch if $t1 and 32 are not equal. The "j" instruction jumps to the calculated address of the showCharsOrSpaces branch. Created the backToBeginning branch. The "sub" instruction subtracts $t0 and $t3 and stores the result in $t3. The "la" instruction loads 0 into the $t3 register.