Update mipsCode
The "addition" instruction adds $t0 and -1 and stores the result in the $t0 register. Created lengthFinder branch in order to find length of input to see in it is too long. The "lb" instruction loads byte into $t1 from $t0. The "addi" instruction adds $t0 and 1 and stores the result in the $t0 register.