Update mipsCode
The "addi" instruction adds $t3 and 1 and stores the result in $t3. The "beq" instruction implements the doMisc branch if $t1 and 10 are equal. The "beq" instruction implements the doMisc branch if $t1 and 0 are equal. The "beq" instruction implements the doMisc branch if $t1 and 32 are equal. The "beq" instruction implements the longInputError branch if $t3 and 5 are equal.