Update mipsCode
The "sub" instruction subtracts $t4 and $t3 and stores the result in the $s1 register. Created greatesPowerFinder branch in order to find greatest power. The "beq" instruction implements the AsciiConvert branch if $s1 and 0 are equal. The "must" instruction multiplies $s0 by $t4. The "mflo" instruction moves the contents of register $t4.